+++ src/ex/compiler.ts
@@ -4,22 +4,16 @@ const workspace = vscode.workspace;
 const window = vscode.window;
 
 import * as cfg from './config';
+import config from '../config';
 import * as work from '../work';
 import * as closure from '../closure';
 import * as util from '../util';
+import * as fs from '../fs';
 import * as externgen from '../externgen';
 
-var latestCompilePath:string = '';
-
-function getSelectedFilePath():string
+function repathToMakeJson(path:string):string
 {
-	return window.activeTextEditor.document.fileName.replace(/\\/g, '/');
-}
-
-function getSelectedMakeJson():string
-{
-	const filename = getSelectedFilePath();
-    return filename.substr(0, filename.lastIndexOf('/')+1) + "make.json";
+    return path.substr(0, path.lastIndexOf('/')+1) + "make.json";
 }
 
 async function generateConfirm(makejson:string, input:string, err:Error):Promise<void>
@@ -38,46 +32,38 @@ export function unload()
 }
 
 export var commands = {
-	'ftpkr.makejson' (){
-		if (!window.activeTextEditor) return;
-		return closure.makeJson(getSelectedMakeJson(), getSelectedFilePath());
+	async 'ftpkr.makejson' (file:vscode.Uri){
+		const selected = await util.fileOrEditorFile(file);
+		await closure.makeJson(repathToMakeJson(selected), selected);
 	},
-	async 'ftpkr.closureCompile' (){
+	async 'ftpkr.closureCompile' (file:vscode.Uri){
 		await cfg.loadTest();
 		await workspace.saveAll();
 		await work.compile.work('ftpkr.closureCompile', async () => {
-			if (!window.activeTextEditor) return;
-			const input = getSelectedFilePath();
-			const makejson = getSelectedMakeJson();
+			const selected = await util.fileOrEditorFile(file).catch(()=>'');
+			var makejson:string = repathToMakeJson(selected);
 			try
 			{
+				if (!await fs.exists(makejson))
+				{
+					const latestPath = util.context.workspaceState.get('latestCompilePath', '');
+					if (!latestPath)
+					{
+						return generateConfirm(makejson, selected, Error('make.json is not found'));
+					}
+					if (!await fs.exists(latestPath))
+					{
+						util.context.workspaceState.update('latestCompilePath', '');
+						return generateConfirm(makejson, selected, Error('make.json is not found'));
+					}
+					makejson = latestPath;
+				}
 				await closure.make(makejson);
-				latestCompilePath = makejson;
+				util.context.workspaceState.update('latestCompilePath', makejson);
 			}
 			catch(err)
 			{
-				if (err.code !== 'ENOENT')
-				{
-					util.log(err);
-					return;
-				}
-				if (latestCompilePath)
-				{
-					return closure.make(latestCompilePath)
-					.catch((err)=>{
-						if (err.code !== 'ENOENT')
-						{
-							util.log(err);
-							return;
-						}
-						latestCompilePath = '';
-						return generateConfirm(makejson, input, err);
-					});
-				}
-				else
-				{
-					return generateConfirm(makejson, input, err);
-				}
+				util.error(err);
 			}
 		});
 	},
@@ -86,12 +72,11 @@ export var commands = {
 		await workspace.saveAll();
 		await work.compile.work('ftpkr.closureCompileAll', () => closure.all());
 	},
-	async 'ftpkr.generateExtern'(){
+	async 'ftpkr.generateExtern'(file:vscode.Uri){
 		await cfg.loadTest();
-		await workspace.saveAll();
-		await work.compile.work('ftpkr.generateExtern', () => {
-			if (!window.activeTextEditor) return;
-			return externgen.gen(getSelectedFilePath());
+		await work.compile.work('ftpkr.generateExtern', async () => {
+			const selected = await util.fileOrEditorFile(file);
+			return externgen.gen(selected);
 		});
 	}
 };