+++ src/index.ts
@@ -10,9 +10,10 @@ const extensions = [
 	require('./ex/compiler')
 ];
 
-function activate(context) {
+function activate(context:vscode.ExtensionContext) {
     console.log('[extension: ftp-kr] activate');
 	fs.setWorkspace(workspace.rootPath.replace(/\\/g, "/"));
+	util.setContext(context);
 
     for(const ex of extensions) ex.load();
 
@@ -22,7 +23,7 @@ function activate(context) {
         {
             let command = ex.commands[p];
             const disposable = vscode.commands.registerCommand(p,(...arg) => command(...arg).catch(util.error));
-            context.subscriptions.push(disposable);
+			context.subscriptions.push(disposable);
         }
     }
 }