+++ src/work.ts
@@ -2,40 +2,26 @@
 class Work
 {
 	promise:Promise<void> = Promise.resolve();
-	commandName:string = '';
-	reserved:Array<string> = [];
+	reserved:string[] = [];
 
 	work(name:string, func:()=>any):Promise<any>
 	{
-		if (this.commandName !== '')
-		{
-			return Promise.reject('ftp-kr is busy: '+this.commandName+' is being proceesed('+name+')');
-		}
 		if (this.reserved.length !== 0)
 		{
 			return Promise.reject('ftp-kr is busy: '+this.reserved[0]+' is being proceesed('+name+')');
 		}
-		this.commandName = name;
-		const prom = this.promise.then(func)
-		.then(v=>{
-			this.commandName = '';
-			return v;
-		});
-		this.promise = prom.catch(err=>{
-			this.commandName = '';
-		});
-		return prom;
+		return this.reserveWork(name, func);
 	}
 	reserveWork(name:string, func:()=>any):Promise<any>
 	{
 		this.reserved.push(name);
 		const prom = this.promise.then(func)
 		.then(v=>{
-			this.reserved.pop();
+			this.reserved.shift();
 			return v;
 		});
 		this.promise = prom.catch(err=>{
-			this.reserved.pop();
+			this.reserved.shift();
 		});
 		return prom;
 	}